pyproject.toml
[tool.poetry]
name = "course"
version = "0.1.0"
description = ""
authors = ["Артем <av.filippov@roseltorg.ru >"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.2"
xlrd = "^2.0.1"
pytest = "^8.2.2"
openpyxl = "^3.1.5"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
yfinance = "^0.2.40"
datetime = "^5.5"


[tool.poetry.group.lint.dependencies]
mypy = "^1.10.1"
black = "^24.4.2"
flake8 = "^7.1.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.black]
line-length = 119
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py

)
'''

[tool.isort]
line_length = 119
